+++ group/general/add-or-remove-user.ps1
@@ -1,0 +1,94 @@
+<#
+  .SYNOPSIS
+  Add/remove users to/from an AzureAD group.
+
+  .DESCRIPTION
+  Add/remove users to/from an AzureAD group.
+
+  .NOTES
+  Permissions: 
+  MS Graph (API)
+  - Group.ReadWrite.All
+  - Directory.ReadWrite.All
+
+  .INPUTS
+  RunbookCustomization: {
+        "Parameters": {
+            "Remove": {
+                "DisplayName": "Add or Remove User",
+                "SelectSimple": {
+                    "Add User as member": false,
+                    "Remove User as member": true
+                }
+            },
+            "GroupId": {
+                "Hide": true
+            },
+            "CallerName": {
+                "Hide": true
+            }
+        }
+    }
+#>
+
+#Requires -Modules @{ModuleName = "RealmJoin.RunbookHelper"; ModuleVersion = "0.6.0" }
+
+param(
+    [Parameter(Mandatory = $true)]
+    [ValidateScript( { Use-RJInterface -Type Graph -Entity Group -DisplayName "Group" } )]
+    [String] $GroupID,
+    [Parameter(Mandatory = $true)]
+    [ValidateScript( { Use-RJInterface -Type Graph -Entity User -DisplayName "User" } )]
+    [String] $UserId,
+    [ValidateScript( { Use-RJInterface -DisplayName "Remove this user" } )]
+    [bool] $Remove = $false,
+    # CallerName is tracked purely for auditing purposes
+    [Parameter(Mandatory = $true)]
+    [string] $CallerName
+)
+
+Connect-RjRbGraph
+
+
+# "Find the user object " 
+$targetUser = Invoke-RjRbRestMethodGraph -Resource "/users/$UserId" -ErrorAction SilentlyContinue
+if (-not $targetUser) {
+    throw ("User '$UserId' not found.")
+}
+
+$targetGroup = Invoke-RjRbRestMethodGraph -Resource "/groups/$GroupID" -ErrorAction SilentlyContinue
+if (-not $targetGroup) {
+    throw ("Group '$GroupID' not found.")
+}
+
+# Work on AzureAD based groups
+if (($targetGroup.GroupTypes -contains "Unified") -or (-not $targetGroup.MailEnabled)) {
+    # Prepare Request
+    $body = @{
+        "@odata.id" = "https://graph.microsoft.com/v1.0/directoryObjects/$UserId"
+    }
+    
+    # "Is user member of the the group?" 
+    if (Invoke-RjRbRestMethodGraph -Resource "/groups/$GroupID/members/$UserID" -ErrorAction SilentlyContinue) {
+        if ($Remove) {
+            Invoke-RjRbRestMethodGraph -Resource "/groups/$GroupID/members/$UserId/`$ref" -Method Delete -Body $body | Out-Null
+            "## '$($targetUser.UserPrincipalName)' is removed from '$($targetGroup.DisplayName)'."
+        }
+        else {    
+            "## User '$($targetUser.UserPrincipalName)' is already a member of '$($targetGroup.DisplayName)'. No action taken."
+        }
+    }
+    else {
+        if ($Remove) {
+            "## User '$($targetUser.UserPrincipalName)' is not a member of '$($targetGroup.DisplayName)'. No action taken."
+        }
+        else {
+            Invoke-RjRbRestMethodGraph -Resource "/groups/$GroupID/members/`$ref" -Method Post -Body $body | Out-Null
+            "## '$($targetUser.UserPrincipalName)' is added to '$($targetGroup.DisplayName)'."  
+        }
+    }
+} 
+else {
+    "## Currently Exchange groups are not supported."
+    throw ("EXO groups not supported")
+}